third_party/WebKit/LayoutTests/fast/dom/HTMLLinkElement/link-sheet-out-of-tree.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var target = document.createElement("link");
shouldBeNull("target.sheet");
</script>
</body>
</html>
